feat: drive abci custom loggers configuration
