+++ packages/dashmate/docker-compose.yml
@@ -72,17 +72,10 @@ services:
       - QUORUM_TYPE=${PLATFORM_DRIVE_ABCI_VALIDATOR_SET_LLMQ_TYPE:?err}
       - NETWORK=${NETWORK:?err}
       - TENDERDASH_P2P_PORT=${PLATFORM_DRIVE_TENDERDASH_P2P_PORT:?err}
-      - ABCI_LOG_STDOUT_DESTINATION=stdout
-      - ABCI_LOG_STDOUT_LEVEL=${PLATFORM_DRIVE_ABCI_LOG_STDOUT_LEVEL:?err}
-      - ABCI_LOG_STDOUT_FORMAT=pretty
-      - ABCI_LOG_STDOUT_COLOR=true
-      - ABCI_LOG_PRETTY_DESTINATION=/var/log/dash/drive/pretty/${PLATFORM_DRIVE_ABCI_LOG_PRETTY_FILE_NAME:?err}
-      - ABCI_LOG_PRETTY_LEVEL=${PLATFORM_DRIVE_ABCI_LOG_PRETTY_FILE_LEVEL:?err}
-      - ABCI_LOG_PRETTY_FORMAT=pretty
-      - ABCI_LOG_JSON_DESTINATION=/var/log/dash/drive/json/${PLATFORM_DRIVE_ABCI_LOG_JSON_FILE_NAME:?err}
-      - ABCI_LOG_JSON_LEVEL=${PLATFORM_DRIVE_ABCI_LOG_JSON_FILE_LEVEL:?err}
-      - ABCI_LOG_JSON_FORMAT=json
     stop_grace_period: 30s
+    env_file:
+      # Logger settings
+      - ${DASHMATE_HOME_DIR:?err}/${CONFIG_NAME:?err}/platform/drive/abci/logger.env.dot
     profiles:
       - platform
