+++ packages/dashmate/src/listr/tasks/startNodeTaskFactory.js
@@ -46,11 +46,15 @@ function startNodeTaskFactory(
 
     // Check Drive log files are created
     if (config.get('platform.enable')) {
-      const prettyFilePath = config.get('platform.drive.abci.log.prettyFile.path');
-      ensureFileMountExists(prettyFilePath, 0o666);
+      const loggers = config.get('platform.drive.abci.logs');
 
-      const jsonFilePath = config.get('platform.drive.abci.log.jsonFile.path');
-      ensureFileMountExists(jsonFilePath, 0o666);
+      for (const logger of Object.values(loggers)) {
+        if (['stdout', 'stderr'].includes(logger.destination)) {
+          continue;
+        }
+
+        ensureFileMountExists(logger.destination, 0o666);
+      }
 
       const tenderdashLogFilePath = config.get('platform.drive.tenderdash.log.path');
       if (tenderdashLogFilePath !== null) {
