+++ packages/dashmate/templates/platform/drive/abci/logger.env.dot
@@ -1,0 +1,7 @@
+{{~ Object.entries(it.platform.drive.abci.logs) :logger }}
+  {{~Object.entries(logger[1]) :setting }}
+    {{? setting[1] !== null }}
+ABCI_LOG_{{=logger[0].toUpperCase()}}_{{=setting[0].toUpperCase()}}={{=setting[1]}}
+    {{?}}
+  {{~}}
+{{~}}
